+++ app/Http/Controllers/GroupController.php
@@ -9,6 +9,7 @@ use App\User;
 use App\UserGroup;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use PhpParser\Node\Expr\New_;
 
 class GroupController extends Controller
@@ -73,6 +74,10 @@ class GroupController extends Controller
     public function show($id)
     {
         $group = Group::find($id);
+
+        if (Gate::denies('can-access-group', $group)){
+            return redirect()->route('home')->with('danger','Vous n\'appartenez pas à ce groupe. Par conséquent vous ne pouvez pas y acceder');
+        }
         $members = UserGroup::where('group_id', $group->id)->get();
         return view('group.show',[
             'group'=>$group,
@@ -89,6 +94,10 @@ class GroupController extends Controller
     public function edit($id)
     {
         $group = Group::find($id);
+
+        if (Gate::denies('can-edit-group', $group)){
+            return redirect()->route('home')->with('danger','Vous n\'êtes pas propriétaire du groupe, vous n\'avez pas accès à cette page');
+        }
 
         $members = UserGroup::where('group_id', $group->id)->where('user_id', '!=', $group->user_id)->get();
         //members dans le 2nd where on exclut du groupe la personne propriétaire et on récupère les autres
@@ -112,6 +121,10 @@ class GroupController extends Controller
     {
         $group = Group::find($id);
 
+        if (Gate::denies('can-edit-group', $group)){
+            return redirect()->route('home')->with('danger','Vous n\'êtes pas propriétaire du groupe, vous ne pouvez pas modifier ses informations');
+        }
+
         if ($request->input('name') != null) {
             $group->name = $request->input('name');
         }
@@ -130,6 +143,11 @@ class GroupController extends Controller
     public function destroy($id)
     {
         $group = Group::find($id);
+
+        if (Gate::denies('can-edit-group', $group)){
+            return redirect()->route('home')->with('danger','Vous n\'êtes pas propriétaire du groupe, vous ne pouvez pas supprimer le groupe');
+        }
+
         $userGroups = UserGroup::where('group_id', $group->id)->get();
         //Try if userGroups array is empty
         if (count($userGroups) > 0) {
@@ -152,14 +170,28 @@ class GroupController extends Controller
     {
         $userGroups = UserGroup::find($id);
 
+        if (Gate::denies('is-member-group', $userGroups->group)){
+            return redirect()->route('home')->with('danger','Vous ne pouvez pas quitter ce groupe');
+        }
+
         $userGroups->delete();
 
         return redirect()->route('group.index')->with('success','Vous avez bien quitter le groupe.');
     }
 
+    /**
+     * @param $id
+     * @param $user_id
+     * @return \Illuminate\Http\RedirectResponse
+     * kick a member from a group
+     */
     public function kick($id, $user_id)
     {
         $group = Group::find($id);
+
+        if (Gate::denies('can-edit-group', $group)){
+            return redirect()->route('home')->with('danger','Vous ne pouvez pas exclure cette personne');
+        }
 
         $member = UserGroup::where('group_id', $id)->where('user_id', $user_id)->get();
 
@@ -171,10 +203,19 @@ class GroupController extends Controller
         return redirect()->route('group.show', $group->id)->with('success','La personne a bien été exclue !');
     }
 
+    /**
+     * @param $id
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     *
+     * display the view with all shares of the group
+     */
     public function share($id){
 
         $group = Group::find($id);
 
+        if (Gate::denies('can-access-group', $group)){
+            return redirect()->route('home')->with('danger','Vous n\'appartenez pas à ce groupe. Par conséquent vous ne pouvez pas y acceder');
+        }
 
         return view('group.share.share',[
             'group'=>$group
